+++ npm/webpack-batteries-included-preprocessor/test/e2e/features.spec.js
@@ -60,10 +60,6 @@ describe('webpack-batteries-included-preprocessor features', () => {
     await runAndEval('node_shim_spec.js')
   })
 
-  it('shims node builtins', async () => {
-    await runAndEval('node_builtins_spec.js')
-  })
-
   it('outputs inline source map', async () => {
     const outputPath = await run('es_features_spec.js')
     const contents = await fs.readFile(outputPath)