+++ packages/data-context/src/sources/ProjectDataSource.ts
@@ -21,7 +21,7 @@ import { toPosix } from '../util/file'
 import type { FilePartsShape } from '@packages/graphql/src/schemaTypes/objectTypes/gql-FileParts'
 import type { ProjectShape } from '../data'
 import type { FindSpecs } from '../actions'
-import { FileExtension, getDefaultSpecFileName } from './migration/utils'
+import { FileExtension, getDefaultSpecFileName } from '../util/files'
 
 type SpecPatterns = {
   specPattern?: string[]