+++ packages/data-context/test/unit/data/ProjectConfigIpc.spec.ts
@@ -38,10 +38,9 @@ describe('ProjectConfigIpc', () => {
   context('forkChildProcess', () => {
     // some of these node versions may not exist, but we want to verify
     // the experimental flags are correctly disabled for future versions
-    const NODE_VERSIONS = ['18.20.4', '20.17.0', '20.19.0', '22.0.0', '22.7.0', '22.11.4', '22.12.0', '22.15.0']
-    const experimentalDetectModuleIntroduced = '22.7.0'
-    const experimentalRequireModuleIntroduced = '22.12.0'
-    const minorPatchExperimentalModuleIntroduced = '>= 20.19.0 < 21.0.0'
+    const NODE_VERSIONS = ['20.5.1', '20.6.0', '20.19.1', '22.15.0']
+
+    const lastVersionWithDeprecatedLoaderOption = '20.5.1'
 
     let projectConfigIpc
     let forkSpy
@@ -59,71 +58,16 @@ describe('ProjectConfigIpc', () => {
 
     context('typescript', () => {
       [...NODE_VERSIONS].forEach((nodeVersion) => {
+        const MOCK_NODE_PATH = `/Users/foo/.nvm/versions/node/v${nodeVersion}/bin/node`
+        const MOCK_NODE_VERSION = nodeVersion
+
         context(`node v${nodeVersion}`, () => {
-          context('ESM', () => {
-            it('passes the correct experimental flags if ESM is being used with typescript', async () => {
-              // @ts-expect-error
-              const projectPath = await scaffoldProject('config-cjs-and-esm/config-with-ts-module')
+          const PROJECTS = ['config-cjs-and-esm/config-with-ts-module', 'config-cjs-and-esm/config-with-module-resolution-bundler', 'config-cjs-and-esm/config-with-js-module', 'config-cjs-and-esm/config-with-cjs']
 
-              const MOCK_NODE_PATH = `/Users/foo/.nvm/versions/node/v${nodeVersion}/bin/node`
-              const MOCK_NODE_VERSION = nodeVersion
-
-              projectConfigIpc = new ProjectConfigIpc(
-                MOCK_NODE_PATH,
-                MOCK_NODE_VERSION,
-                projectPath,
-                'cypress.config.js',
-                false,
-                (error) => {},
-                () => {},
-                () => {},
-              )
-
-              expect(forkSpy).to.have.been.calledWith(sinon.match.string, sinon.match.array, sinon.match({
-                env: {
-                  NODE_OPTIONS: sinon.match('--experimental-specifier-resolution=node --loader'),
-                },
-              }))
-
-              if (semver.gte(nodeVersion, experimentalDetectModuleIntroduced) || semver.satisfies(nodeVersion, minorPatchExperimentalModuleIntroduced)) {
-                expect(forkSpy).to.have.been.calledWith(sinon.match.string, sinon.match.array, sinon.match({
-                  env: {
-                    NODE_OPTIONS: sinon.match('--no-experimental-detect-module'),
-                  },
-                }))
-              }
-
-              if (semver.gte(nodeVersion, experimentalRequireModuleIntroduced) || semver.satisfies(nodeVersion, minorPatchExperimentalModuleIntroduced)) {
-                expect(forkSpy).to.have.been.calledWith(sinon.match.string, sinon.match.array, sinon.match({
-                  env: {
-                    NODE_OPTIONS: sinon.match('--no-experimental-require-module'),
-                  },
-                }))
-              }
-
-              if (semver.eq(nodeVersion, '22.0.0')) {
-                expect(forkSpy).to.not.have.been.calledWith(sinon.match.string, sinon.match.array, sinon.match({
-                  env: {
-                    NODE_OPTIONS: sinon.match('--no-experimental-detect-module'),
-                  },
-                }))
-
-                expect(forkSpy).to.not.have.been.calledWith(sinon.match.string, sinon.match.array, sinon.match({
-                  env: {
-                    NODE_OPTIONS: sinon.match('--no-experimental-require-module'),
-                  },
-                }))
-              }
-            })
-          })
-
-          context('CommonJS', () => {
-            it('uses the ts_node commonjs loader if CommonJS is being used with typescript', async () => {
-              // @ts-expect-error
-              const projectPath = await scaffoldProject('config-cjs-and-esm/config-with-module-resolution-bundler')
-
-              const MOCK_NODE_PATH = `/Users/foo/.nvm/versions/node/v${nodeVersion}/bin/node`
-              const MOCK_NODE_VERSION = nodeVersion
+          PROJECTS.forEach((project) => {
+            it(`${project}: tsx generic loader (esm/commonjs/typescript)`, async () => {
+              // @ts-expect-error ignoring due to nested directories in the system-test project directory not being included in the type.
+              const projectPath = await scaffoldProject(project)
 
               projectConfigIpc = new ProjectConfigIpc(
                 MOCK_NODE_PATH,
@@ -136,11 +80,38 @@ describe('ProjectConfigIpc', () => {
                 () => {},
               )
 
-              expect(forkSpy).to.have.been.calledWith(sinon.match.string, sinon.match.array, sinon.match({
-                env: {
-                  NODE_OPTIONS: sinon.match('--require'),
-                },
-              }))
+              // make sure that we use tsx for every file, regardless of typescript, esm, or commonjs
+              if (semver.lte(nodeVersion, lastVersionWithDeprecatedLoaderOption)) {
+                // For node 20.5.1 and down, we need use the --loader flag
+                expect(forkSpy).to.have.been.calledWith(sinon.match.string, sinon.match.array, sinon.match({
+                  env: {
+                    NODE_OPTIONS: sinon.match(/--loader .*cypress\/node_modules\/tsx\/dist\/loader.mjs/),
+                  },
+                }))
+              } else {
+                // For node 20.6.0 and up, we need use the --import flag
+                expect(forkSpy).to.have.been.calledWith(sinon.match.string, sinon.match.array, sinon.match({
+                  env: {
+                    NODE_OPTIONS: sinon.match(/--import .*cypress\/node_modules\/tsx\/dist\/loader.mjs/),
+                  },
+                }))
+              }
+
+              if (project.includes('config-with-ts-module') || project.includes('config-with-module-resolution-bundler')) {
+                // these projects have typescript installed and have a tsconfig, so the TSX_TSCONFIG_PATH should be set to the project path
+                expect(forkSpy).to.have.been.calledWith(sinon.match.string, sinon.match.array, sinon.match({
+                  env: {
+                    TSX_TSCONFIG_PATH: sinon.match(`/cy-projects/${project}/tsconfig.json`),
+                  },
+                }))
+              } else {
+                // non typescript projects that do NOT have a tsconfig, so the TSX_TSCONFIG_PATH should be undefined
+                expect(forkSpy).to.have.been.calledWith(sinon.match.string, sinon.match.array, sinon.match({
+                  env: {
+                    TSX_TSCONFIG_PATH: undefined,
+                  },
+                }))
+              }
             })
           })
         })