+++ packages/launchpad/src/migration/fragments/MigrationListItem.vue
@@ -1,7 +1,0 @@
-<template>
-  <li class="mb-[4px]">
-    <span class="text-gray-600">
-      <slot />
-    </span>
-  </li>
-</template>