+++ system-tests/__snapshots__/retries_spec.ts.js
@@ -60,67 +60,6 @@ exports['retries / supports retries'] = `
 
 `
 
-exports['retries / warns about retries plugin'] = `
-We've detected that the incompatible plugin cypress-plugin-retries is installed at: node_modules/cypress-plugin-retries
-
-Test retries is now natively supported in Cypress version 5.0.0.
-
-Remove the plugin from your dependencies to silence this warning.
-
-https://on.cypress.io/test-retries
-
-
-====================================================================================================
-
-  (Run Starting)
-
-  ┌────────────────────────────────────────────────────────────────────────────────────────────────┐
-  │ Cypress:    1.2.3                                                                              │
-  │ Browser:    FooBrowser 88                                                                      │
-  │ Specs:      1 found (main.spec.cy.js)                                                          │
-  │ Searched:   cypress/e2e/main.spec.cy.js                                                        │
-  └────────────────────────────────────────────────────────────────────────────────────────────────┘
-
-
-────────────────────────────────────────────────────────────────────────────────────────────────────
-                                                                                                    
-  Running:  main.spec.cy.js                                                                 (1 of 1)
-
-
-  ✓ foo
-
-  1 passing
-
-
-  (Results)
-
-  ┌────────────────────────────────────────────────────────────────────────────────────────────────┐
-  │ Tests:        1                                                                                │
-  │ Passing:      1                                                                                │
-  │ Failing:      0                                                                                │
-  │ Pending:      0                                                                                │
-  │ Skipped:      0                                                                                │
-  │ Screenshots:  0                                                                                │
-  │ Video:        false                                                                            │
-  │ Duration:     X seconds                                                                        │
-  │ Spec Ran:     main.spec.cy.js                                                                  │
-  └────────────────────────────────────────────────────────────────────────────────────────────────┘
-
-
-====================================================================================================
-
-  (Run Finished)
-
-
-       Spec                                              Tests  Passing  Failing  Pending  Skipped  
-  ┌────────────────────────────────────────────────────────────────────────────────────────────────┐
-  │ ✔  main.spec.cy.js                          XX:XX        1        1        -        -        - │
-  └────────────────────────────────────────────────────────────────────────────────────────────────┘
-    ✔  All specs passed!                        XX:XX        1        1        -        -        -  
-
-
-`
-
 exports['retries / supports retries (chrome)'] = `
 
 ====================================================================================================