+++ system-tests/projects/invalid-root-level-config/invalid-componentFolder-config.js
@@ -1,4 +1,0 @@
-module.exports = {
-  componentFolder: 'src/',
-  e2e: {},
-}