+++ system-tests/test/plugins_spec.js
@@ -207,6 +207,7 @@ describe('e2e plugins', function () {
 
     it('supports node builtins', function () {
       return systemTests.exec(this, {
+        project: 'node-builtins',
         spec: 'node_builtins.cy.js',
       })
     })